[CHERRY-PICK] Adding '4b7d35377d9334e20369750674ba18626ca63a9c' to 'CHERRY_PICK_main_2024-03-14_15.44.12' using the python script version 'v1.0.8'.

+++ PresentationLayer/ESTABLISHMENT/LEAVES/Transactions/DirectLeaveApproval.aspx.cs
@@ -437,7 +437,7 @@ public partial class ESTABLISHMENT_LEAVES_Transactions_RegistrarLeaveApproval : 
                     checkcount = checkcount + 1;
                     LETRNO = Convert.ToInt32(chkSelect.ToolTip);
                     DataSet ds = objApp.GetBalanceforDirectLeaveApproval(LETRNO, IDNO, LeaveNo);
-                    int bal = Convert.ToInt32(ds.Tables[0].Rows[0]["CLBal"].ToString());
+                    double bal = Convert.ToDouble(ds.Tables[0].Rows[0]["CLBal"].ToString());
 
                     if (bal >= noOfDays || Status.Equals("R"))
                     {
